+++ lib/firestore_helper.dart
@@ -6,7 +6,6 @@ import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firestore_helper/src/log_type.dart';
 import 'package:firestore_helper/src/logging_service.dart';
 import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
 import 'package:logger/logger.dart';
 
 part 'src/firestore_helper.dart';